# openapi/authz-api.yaml
openapi: 3.0.3
info:
  title: Busibox AuthZ API
  description: |
    Authorization and authentication service for Busibox platform.
    
    Provides:
    - OAuth2 token issuance (client_credentials and token-exchange grants)
    - RBAC management (roles, users, permissions)
    - Audit logging
    - JWKS endpoint for token validation
    
    **Authentication**: OAuth client credentials or admin token required for most endpoints.
  version: 1.0.0
  contact:
    name: Busibox Team
servers:
  - url: http://10.96.200.210:8010
    description: Production AuthZ Service
  - url: http://10.96.201.210:8010
    description: Test AuthZ Service
...
paths:
...
  /admin/roles:
    get:
      tags: [Admin]
      summary: List all roles
      description: Returns all roles in the system. Requires admin authentication.
      operationId: listRoles
      security:
        - AdminToken: []
        - OAuthClient: []
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Admin]
      summary: Create a new role
      description: Creates a new role. Requires admin authentication.
      operationId: createRole
      security:
        - AdminToken: []
        - OAuthClient: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/RoleCreate'
                - $ref: '#/components/schemas/OAuthClientAuth'
      responses:
        '200':
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/roles/{role_id}:
    get:
      tags: [Admin]
      summary: Get a specific role
      description: Returns details of a specific role by ID. Requires admin authentication.
      operationId: getRole
      security:
        - AdminToken: []
        - OAuthClient: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Role details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Admin]
      summary: Update a role
      description: Updates an existing role. Requires admin authentication.
      operationId: updateRole
      security:
        - AdminToken: []
        - OAuthClient: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/RoleUpdate'
                - $ref: '#/components/schemas/OAuthClientAuth'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Admin]
      summary: Delete a role
      description: Deletes a role and all associated user-role bindings. Requires admin authentication.
      operationId: deleteRole
      security:
        - AdminToken: []
        - OAuthClient: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Role deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  deleted:
                    type: boolean
                    example: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/user-roles:
    post:
      tags: [Admin]
      summary: Add user-role binding
      description: Assigns a role to a user. Requires admin authentication.
      operationId: addUserRole
      security:
        - AdminToken: []
        - OAuthClient: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UserRoleBinding'
                - $ref: '#/components/schemas/OAuthClientAuth'
      responses:
        '200':
          description: User-role binding created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRoleBindingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [Admin]
      summary: Remove user-role binding
      description: Removes a role from a user. Requires admin authentication.
      operationId: removeUserRole
      security:
        - AdminToken: []
        - OAuthClient: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UserRoleBinding'
                - $ref: '#/components/schemas/OAuthClientAuth'
      responses:
        '200':
          description: User-role binding removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  deleted:
                    type: boolean
                    example: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{user_id}/roles:
    get:
      tags: [Admin]
      summary: Get user roles
      description: Returns all roles assigned to a specific user. Requires admin authentication.
      operationId: getUserRoles
      security:
        - AdminToken: []
        - OAuthClient: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of user roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/oauth-clients:
    get:
      tags: [Admin]
      summary: List OAuth clients
      description: Returns all OAuth clients. Requires admin authentication.
      operationId: listOAuthClients
      security:
        - AdminToken: []
        - OAuthClient: []
      responses:
        '200':
          description: List of OAuth clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OAuthClientResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Admin]
      summary: Create OAuth client
      description: Creates a new OAuth client. Requires admin authentication.
      operationId: createOAuthClient
      security:
        - AdminToken: []
        - OAuthClient: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthClientCreate'
      responses:
        '201':
          description: OAuth client created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthClientResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    AdminToken:
      type: http
      scheme: bearer
      description: Admin token for RBAC operations (AUTHZ_ADMIN_TOKEN)

    OAuthClient:
      type: oauth2
      description: OAuth client credentials flow
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
          scopes: {}
  schemas:
...
    RoleCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: Editor
        description:
          type: string
          example: Can edit documents
        scopes:
          type: array
          description: OAuth2 scopes granted by this role
          items:
            type: string
          example: [ingest.read, ingest.write, search.read]
    RoleUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        scopes:
          type: array
          description: OAuth2 scopes granted by this role
          items:
            type: string
          example: [ingest.read, ingest.write, search.read]
    RoleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Editor
        description:
          type: string
          example: Can edit documents
        scopes:
          type: array
          description: OAuth2 scopes granted by this role
          items:
            type: string
          example: [ingest.read, ingest.write, search.read]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserRoleBinding:
      type: object
      required:
        - user_id
        - role_id
      properties:
        user_id:
          type: string
          format: uuid
        role_id:
          type: string
          format: uuid
    UserRoleBindingResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        role_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    OAuthClientCreate:
      type: object
      required:
        - client_id
        - client_secret
      properties:
        client_id:
          type: string
          description: Client ID for the new OAuth client
        client_secret:
          type: string
          description: Client secret (will be hashed)
        allowed_audiences:
          type: array
          items:
            type: string
          example: [ingest-api, search-api, agent-api]
        allowed_scopes:
          type: array
          items:
            type: string
          example: [ingest.read, ingest.write, search.read]
        auth_client_id:
          type: string
          description: Client ID for authentication (if not using admin token)
        auth_client_secret:
          type: string
          description: Client secret for authentication (if not using admin token)
    OAuthClientResponse:
      type: object
      properties:
        client_id:
          type: string
        allowed_audiences:
          type: array
          items:
            type: string
        allowed_scopes:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    OAuthClientAuth:
      type: object
      description: OAuth client credentials for authentication (alternative to admin token)
      properties:
        client_id:
          type: string
        client_secret:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
        detail:
          type: string
          description: Error details

    # ========================================================================
    # Keystore Schemas (Envelope Encryption)
    # ========================================================================
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'